refactor(anima_train): 抽 timestep_sampling / noise / loss_weighting 到 training/
ADR 0003 PR-A 第 4 步：3 个数学小模块，单文件多 fn 模式（不开 plugin
subfolder，每个 scheme 是同签名 ~30 行函数）：

- training/timestep_sampling.py — sample_t (logit_normal/uniform/mode/...)
- training/noise.py             — make_noise (offset + pyramid)
- training/loss_weighting.py    — compute_loss_weight (min_snr/cosmap/...)

anima_train.py 行数 1042 → 930；migration test 全绿。
