+++ runtime/anima_train.py
@@ -106,6 +106,9 @@ from training.cli import (  # noqa: E402
     parse_args,
     prompt_for_args,
 )
+from training.timestep_sampling import sample_t  # noqa: E402
+from training.noise import make_noise  # noqa: E402
+from training.loss_weighting import compute_loss_weight  # noqa: E402
 
 
 # 模型加载（load_anima_model / load_vae / load_text_encoders / ensure_models_namespace）
@@ -122,134 +125,6 @@ from training.cli import (  # noqa: E402
 
 
 # 训练时推理 sample_image 已搬到 training.sampling（ADR 0003 PR-A）。
-
-
-# ============================================================================
-# 训练辅助
-# ============================================================================
-
-def sample_t(bs, device, mode: str = "logit_normal", shift: float = 3.0) -> torch.Tensor:
-    """采样 Flow Matching 时间步 t ∈ (0, 1)。
-
-    mode:
-      logit_normal      — SD3/Anima 默认，偏向中间 t；shift>1 推向高噪声端
-      uniform           — 均匀采样，对细节端和结构端覆盖更均衡
-      logit_normal_low  — logit-normal 反向 shift，偏向低噪声/细节端
-      mode              — SD3 mode-distribution，集中在某个 sigma 附近
-    """
-    mode = (mode or "logit_normal").lower()
-    u = torch.sigmoid(torch.randn(bs, device=device))
-
-    if mode == "uniform":
-        return torch.rand(bs, device=device).clamp(1e-4, 1 - 1e-4)
-
-    if mode == "logit_normal_low":
-        s = max(float(shift), 1e-4)
-        u = (u * (1.0 / s)) / (1 + (1.0 / s - 1) * u)
-        return u.clamp(1e-4, 1 - 1e-4)
-
-    if mode == "mode":
-        s = float(shift)
-        u = 1 - u - s * (torch.cos(torch.pi * 0.5 * u) ** 2 - 1 + u)
-        return u.clamp(1e-4, 1 - 1e-4)
-
-    # logit_normal（默认）+ shift
-    s = float(shift)
-    u = (u * s) / (1 + (s - 1) * u)
-    return u.clamp(1e-4, 1 - 1e-4)
-
-
-def make_noise(
-    latents: torch.Tensor,
-    noise_offset: float = 0.0,
-    pyramid_iters: int = 0,
-    pyramid_discount: float = 0.35,
-) -> torch.Tensor:
-    """生成训练噪声，可叠加低频扰动。
-
-    noise_offset   — 给每样本/通道加低频偏移，缓解亮度均值偏差（SDXL 思路）
-    pyramid_iters  — 叠加多尺度低频噪声，帮助模型快速学习全局光照/构图；
-                     bilinear 插值避免 nearest 的块状结构干扰
-    """
-    noise = torch.randn_like(latents)
-
-    if noise_offset > 0:
-        shape = list(latents.shape)
-        for ax in range(2, latents.ndim):
-            shape[ax] = 1
-        offset = torch.randn(*shape, device=latents.device, dtype=latents.dtype)
-        noise = noise + noise_offset * offset
-
-    if pyramid_iters > 0:
-        try:
-            spatial = list(latents.shape[-2:])
-            cur = noise.clone()
-            for i in range(pyramid_iters):
-                r = 2 ** (i + 1)
-                sh, sw = max(spatial[0] // r, 1), max(spatial[1] // r, 1)
-                if latents.ndim == 5:
-                    extra = torch.randn(
-                        latents.shape[0], latents.shape[1], latents.shape[2], sh, sw,
-                        device=latents.device, dtype=latents.dtype,
-                    )
-                    extra = F.interpolate(
-                        extra.flatten(0, 1), size=spatial, mode="bilinear", align_corners=False,
-                    ).view(latents.shape[0], latents.shape[1], latents.shape[2], *spatial)
-                else:
-                    extra = torch.randn(latents.shape[0], latents.shape[1], sh, sw,
-                                        device=latents.device, dtype=latents.dtype)
-                    extra = F.interpolate(extra, size=spatial, mode="bilinear", align_corners=False)
-                cur = cur + extra * (pyramid_discount ** (i + 1))
-                if min(sh, sw) <= 1:
-                    break
-            noise = cur / cur.std().clamp(min=1e-6)
-        except Exception as exc:
-            logger.warning(f"pyramid_noise 失败，回退标准噪声: {exc}")
-
-    return noise
-
-
-def compute_loss_weight(
-    t: torch.Tensor,
-    scheme: str = "none",
-    min_snr_gamma: float = 5.0,
-    weight_cap_ratio: float = 0.0,
-) -> torch.Tensor:
-    """返回每样本 loss 权重 (B,)，Flow Matching CONST 调度下：SNR(t) = ((1-t)/t)^2。
-
-    scheme:
-      none          — 全 1，与原始行为一致
-      min_snr       — w = min(gamma/SNR, 1)，下调高 SNR 简单步（推荐基础款）
-      detail_inv_t  — w = 1/t clamp [1,5]，温和细节强化，小 batch + Prodigy 友好
-      cosmap        — SD3 cosmap weighting，中间 t 更均匀（max/min ≈ 1.81×）
-
-    weight_cap_ratio — batch 内 max/min 比上限（0=禁用），防单样本主导破坏 Prodigy d 估计
-    """
-    scheme = (scheme or "none").lower()
-    if scheme == "none":
-        return torch.ones_like(t)
-
-    eps = 1e-4
-    t_c = t.clamp(eps, 1 - eps)
-
-    if scheme == "min_snr":
-        snr = ((1 - t_c) / t_c) ** 2
-        w = torch.minimum(torch.tensor(float(min_snr_gamma), device=t.device) / snr, torch.ones_like(t_c))
-    elif scheme == "detail_inv_t":
-        w = (1.0 / t_c).clamp(min=1.0, max=5.0)
-    elif scheme == "cosmap":
-        bot = (1 - 2 * t_c + 2 * t_c ** 2).clamp(min=eps)
-        w = 2.0 / (math.pi * bot)
-    else:
-        return torch.ones_like(t)
-
-    if weight_cap_ratio and weight_cap_ratio > 1.0:
-        w_min = w.min().clamp(min=eps)
-        w = w.clamp(max=w_min * float(weight_cap_ratio))
-
-    return w
-
-
 
 
 # ============================================================================
